Align against javascript callbacks in GNU social
Force a reload when a new notice has been posted. This should probably be
a redirect to whatever page the new notice can be seen on (either itself
or a relevant feed).

This is implemented since GNU social don't respond with 303 See Other on
AJAX calls, and the newer GNU social code makes sure the call is done with
ajax=1 if the call is made with AJAX.

Requires GNU social commit 7c373edf4d86c704320743614c696a44ef363858 from
Jun 14 2015.